Move `GenesisAccount`  struct and `FeeCalculator` trait into fp-evm
