+++ frame/dynamic-fee/Cargo.toml
@@ -20,7 +20,7 @@ sp-std = { version = "4.0.0", git = "https://github.com/paritytech/substrate", b
 frame-support = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
 frame-system = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
 
-pallet-evm = { version = "6.0.0-dev", path = "../evm", default-features = false }
+fp-evm = { version = "3.0.0-dev", path = "../../primitives/evm", default-features = false }
 
 [dev-dependencies]
 pallet-timestamp = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
@@ -40,5 +40,5 @@ std = [
 	"frame-system/std",
 	"frame-support/std",
 
-	"pallet-evm/std",
+	"fp-evm/std",
 ]
