+++ template/node/Cargo.toml
@@ -60,11 +60,11 @@ fc-mapping-sync = { path = "../../client/mapping-sync" }
 fc-rpc = { path = "../../client/rpc" }
 fc-rpc-core = { path = "../../client/rpc-core" }
 fp-consensus = { path = "../../primitives/consensus" }
+fp-evm = { path = "../../primitives/evm" }
 fp-rpc = { path = "../../primitives/rpc" }
 fp-storage = { path = "../../primitives/storage" }
 
 pallet-dynamic-fee = { path = "../../frame/dynamic-fee" }
-pallet-evm = { path = "../../frame/evm" }
 
 frontier-template-runtime = { path = "../runtime", default-features = false, features = ["std"] }
