+++ template/node/src/chain_spec.rs
@@ -173,7 +173,7 @@ fn testnet_genesis(
 					// Using the full hex key, truncating to the first 20 bytes (the first 40 hex chars)
 					H160::from_str("d43593c715fdd31c61141abd04a99fd6822c8558")
 						.expect("internal H160 is valid; qed"),
-					pallet_evm::GenesisAccount {
+					fp_evm::GenesisAccount {
 						balance: U256::from_str("0xffffffffffffffffffffffffffffffff")
 							.expect("internal U256 is valid; qed"),
 						code: Default::default(),
@@ -185,7 +185,7 @@ fn testnet_genesis(
 					// H160 address of CI test runner account
 					H160::from_str("6be02d1d3665660d22ff9624b7be0551ee1ac91b")
 						.expect("internal H160 is valid; qed"),
-					pallet_evm::GenesisAccount {
+					fp_evm::GenesisAccount {
 						balance: U256::from_str("0xffffffffffffffffffffffffffffffff")
 							.expect("internal U256 is valid; qed"),
 						code: Default::default(),
